apps/requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
plotly==5.17.0
google-generativeai==0.3.0
python-dotenv==1.0.0
